Creating mongodb docker-compose definitions Adding spring-data-mongo lib Refactoring classes to be in a repository in-memory
