+++ src/main/java/com/dis/bot/commands/age/character/ThirteenAgeRPGCharacterCreateCommand.java
@@ -1,7 +1,7 @@
 package com.dis.bot.commands.age.character;
 
 import com.dis.bot.commands.SlashCommand;
-import com.dis.bot.repository.age.AgeCharacters;
+import com.dis.bot.repository.inmemory.age.AgeCharacters;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Mono;
