+++ src/main/java/com/dis/bot/commands/character/CharacterCreateCommand.java
@@ -1,7 +1,7 @@
 package com.dis.bot.commands.character;
 
 import com.dis.bot.commands.SlashCommand;
-import com.dis.bot.repository.Characters;
+import com.dis.bot.repository.inmemory.Characters;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Mono;
