+++ src/main/java/com/dis/bot/service/age/PhysicalDefenseService.java
@@ -1,7 +1,7 @@
 package com.dis.bot.service.age;
 
 import com.dis.bot.pojo.character.RPGCharacter;
-import com.dis.bot.repository.age.AgeCharacters;
+import com.dis.bot.repository.inmemory.age.AgeCharacters;
 import org.springframework.stereotype.Service;
 
 import static com.dis.bot.tool.D20Roll.rollD20WithBonus;
